Change code deprecation
I'm scrapping the use of Kociemba's algorithm as it seems more complicated than it does at first glance. There's also great potential for performance issues occuring in case the host machine does not have enough memory for performing search and pruning on roughly 2 million cases, which is an operation that occurs twice in the algorithm. If anyone knows how to reason through the implementation of the algoirthm, do tell.