Add console clearing capability
This program is expected to be used interactively only. Therefore, we add behavior to the program such that the console is cleared every time the user interacts with the menu. That way, any messages printed because of user input appear at the top and the menu appears again for the user to enter their next interaction, which keeps the UI clean.

NOTE: I originally implemented this behavior in the CSCI 112 project and copied it over to here.